moved to cart and attendee class
